chore: removes unused docker compose file fix: Fixes a bug where the build doesn't check if it can't delete anything before it does

+++ app/src/scripts/build.js
@@ -14,8 +14,9 @@ const binPath = path.resolve('bin');
 let config = require(path.resolve('src/config/webpack.prod'));
 
 // Clean out the old files first
-// @TODO: this break if there is no publicDirectory, aka the first build
-fs.rmSync(binPath + '/' + server.publicDirectory, {recursive: true});
+if (fs.existsSync(binPath + '/' + server.publicDirectory)) {
+    fs.rmSync(binPath + '/' + server.publicDirectory, {recursive: true});
+}
 promisify(fs.readdir)(binPath).then(files => {
     for (const file of files) {
         fs.unlinkSync(binPath + '/' + file);
